pyproject.toml
[project]
name = "mailstrom"
version = "0.0.1"
description = "Pulumi-built infrastructure for Thundermail"
requires-python = ">=3.12"
dynamic = ["dependencies"]

[project.urls]
repository = "https://github.com/thunderbird/mailstrom.git"
issues = "https://github.com/thunderbird/mailstrom/issues"

[tool.setuptools.dynamic]
dependencies = { file = ["pulumi/requirements.txt"] }

[project.optional-dependencies]
dev = [
    "ruff",
    "python-dotenv==1.0.1",
    "pytest==8.3.4",
    "locust==2.36.2",
]

[pytest]
log_cli = 1
